backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
torch
torchvision
transformers
peft
qwen_vl_utils
